+++ R/as_hce.default.R
@@ -11,8 +11,10 @@
 #' dat <- KHCE
 #' class(dat) <- "moo" # non-existent class
 #' as_hce(dat) # tries to convert to an hce object
+#' ## It still works because the inheritance converted it to a data frame
 as_hce.default <- function(x, ...){
-  ## Add inheritance to data frame
+  ##Sometimes, it is not possible to convert an object directly to a data frame. 
+  ##However, you can often add inheritance to an object so that it behaves like a data frame. 
   x <- base::structure(x, class = c(class(x), "data.frame"))
   ## call the data frame method of the function
   as_hce(x)